test(qa): check in voice call plugin fixture
